studyspark-backend/src/Summarizer/requirements.txt
fastapi
uvicorn
python-multipart
transformers
langchain
chromadb
textstat
langdetect
textblob
sentence-transformers
PyMuPDF==1.22.5
python-docx
